requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.9.0
